fix: export demo notificaiton setting component to its own file
